refactor(dto): move dto classes into dto package
