+++ src/main/kotlin/lotto/view/dto/LottoTicketDto.kt
@@ -1,4 +1,4 @@
-package lotto.view
+package lotto.view.dto
 
 data class LottoTicketDto(val numbers: List<Int>) {
 }
